+++ internal/provider/data_source_ise_network_access_condition.go
@@ -25,6 +25,7 @@ import (
 	"fmt"
 	"net/url"
 
+	"github.com/CiscoDevNet/terraform-provider-ise/internal/provider/helpers"
 	"github.com/hashicorp/terraform-plugin-framework-validators/datasourcevalidator"
 	"github.com/hashicorp/terraform-plugin-framework/datasource"
 	"github.com/hashicorp/terraform-plugin-framework/datasource/schema"
@@ -106,6 +107,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 			},
 			"operator": schema.StringAttribute{
 				MarkdownDescription: "Equality operator",
+				CustomType:          helpers.OperatorType{},
 				Computed:            true,
 			},
 			"children": schema.ListNestedAttribute{
@@ -151,6 +153,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 						},
 						"operator": schema.StringAttribute{
 							MarkdownDescription: "Equality operator",
+							CustomType:          helpers.OperatorType{},
 							Computed:            true,
 						},
 						"children": schema.ListNestedAttribute{
@@ -196,6 +199,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 									},
 									"operator": schema.StringAttribute{
 										MarkdownDescription: "Equality operator",
+										CustomType:          helpers.OperatorType{},
 										Computed:            true,
 									},
 									"children": schema.ListNestedAttribute{
@@ -233,6 +237,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 												},
 												"operator": schema.StringAttribute{
 													MarkdownDescription: "Equality operator",
+													CustomType:          helpers.OperatorType{},
 													Computed:            true,
 												},
 												"children": schema.ListNestedAttribute{
@@ -270,6 +275,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 															},
 															"operator": schema.StringAttribute{
 																MarkdownDescription: "Equality operator",
+																CustomType:          helpers.OperatorType{},
 																Computed:            true,
 															},
 															"children": schema.ListNestedAttribute{
@@ -307,6 +313,7 @@ func (d *NetworkAccessConditionDataSource) Schema(ctx context.Context, req datas
 																		},
 																		"operator": schema.StringAttribute{
 																			MarkdownDescription: "Equality operator",
+																			CustomType:          helpers.OperatorType{},
 																			Computed:            true,
 																		},
 																		"children": schema.ListNestedAttribute{