+++ internal/provider/model_ise_active_directory_groups_by_domain.go
@@ -25,6 +25,7 @@ import (
 	"fmt"
 	"net/url"
 
+	"github.com/CiscoDevNet/terraform-provider-ise/internal/provider/helpers"
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/tidwall/gjson"
 	"github.com/tidwall/sjson"
@@ -171,6 +172,7 @@ func (data *ActiveDirectoryGroupsByDomain) updateFromBody(ctx context.Context, r
 	for i := range data.Groups {
 		keys := [...]string{"name", "sid", "type"}
 		keyValues := [...]string{data.Groups[i].Name.ValueString(), data.Groups[i].Sid.ValueString(), data.Groups[i].Type.ValueString()}
+		keyNormalize := [...]bool{false, false, false}
 
 		var r gjson.Result
 		parentItems := res.Get("ERSActiveDirectoryGroups.groups").Array()
@@ -178,7 +180,7 @@ func (data *ActiveDirectoryGroupsByDomain) updateFromBody(ctx context.Context, r
 		for _, v := range parentItems {
 			found := false
 			for ik := range keys {
-				if v.Get(keys[ik]).String() == keyValues[ik] {
+				if helpers.MatchKey(v.Get(keys[ik]).String(), keyValues[ik], keyNormalize[ik]) {
 					found = true
 					continue
 				}