+++ internal/provider/model_ise_allowed_protocols.go
@@ -371,7 +371,7 @@ func (data *AllowedProtocols) fromBody(ctx context.Context, res gjson.Result) {
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("AllowedProtocols.description"); value.Exists() && value.Type != gjson.Null {
+	if value := res.Get("AllowedProtocols.description"); value.Exists() && value.Type != gjson.Null && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()
@@ -757,7 +757,7 @@ func (data *AllowedProtocols) updateFromBody(ctx context.Context, res gjson.Resu
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("AllowedProtocols.description"); value.Exists() && !data.Description.IsNull() {
+	if value := res.Get("AllowedProtocols.description"); value.Exists() && !data.Description.IsNull() && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()