+++ internal/provider/model_ise_allowed_protocols_tacacs.go
@@ -97,7 +97,7 @@ func (data *AllowedProtocolsTACACS) fromBody(ctx context.Context, res gjson.Resu
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("AllowedProtocols.description"); value.Exists() && value.Type != gjson.Null {
+	if value := res.Get("AllowedProtocols.description"); value.Exists() && value.Type != gjson.Null && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()
@@ -128,7 +128,7 @@ func (data *AllowedProtocolsTACACS) updateFromBody(ctx context.Context, res gjso
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("AllowedProtocols.description"); value.Exists() && !data.Description.IsNull() {
+	if value := res.Get("AllowedProtocols.description"); value.Exists() && !data.Description.IsNull() && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()