+++ internal/provider/model_ise_certificate_authentication_profile.go
@@ -91,7 +91,7 @@ func (data *CertificateAuthenticationProfile) fromBody(ctx context.Context, res 
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("CertificateProfile.description"); value.Exists() && value.Type != gjson.Null {
+	if value := res.Get("CertificateProfile.description"); value.Exists() && value.Type != gjson.Null && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()
@@ -132,7 +132,7 @@ func (data *CertificateAuthenticationProfile) updateFromBody(ctx context.Context
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("CertificateProfile.description"); value.Exists() && !data.Description.IsNull() {
+	if value := res.Get("CertificateProfile.description"); value.Exists() && !data.Description.IsNull() && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()