+++ internal/provider/model_ise_trustsec_ip_to_sgt_mapping.go
@@ -95,7 +95,7 @@ func (data *TrustSecIPToSGTMapping) fromBody(ctx context.Context, res gjson.Resu
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("SGMapping.description"); value.Exists() && value.Type != gjson.Null {
+	if value := res.Get("SGMapping.description"); value.Exists() && value.Type != gjson.Null && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()
@@ -141,7 +141,7 @@ func (data *TrustSecIPToSGTMapping) updateFromBody(ctx context.Context, res gjso
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("SGMapping.description"); value.Exists() && !data.Description.IsNull() {
+	if value := res.Get("SGMapping.description"); value.Exists() && !data.Description.IsNull() && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()