+++ internal/provider/model_ise_trustsec_security_group_acl.go
@@ -83,7 +83,7 @@ func (data *TrustSecSecurityGroupACL) fromBody(ctx context.Context, res gjson.Re
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("Sgacl.description"); value.Exists() && value.Type != gjson.Null {
+	if value := res.Get("Sgacl.description"); value.Exists() && value.Type != gjson.Null && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()
@@ -109,7 +109,7 @@ func (data *TrustSecSecurityGroupACL) updateFromBody(ctx context.Context, res gj
 	} else {
 		data.Name = types.StringNull()
 	}
-	if value := res.Get("Sgacl.description"); value.Exists() && !data.Description.IsNull() {
+	if value := res.Get("Sgacl.description"); value.Exists() && !data.Description.IsNull() && value.String() != "" {
 		data.Description = types.StringValue(value.String())
 	} else {
 		data.Description = types.StringNull()